Jatj/bug fix

See merge request cpsc319-2018w2/jabc/oecm/jabc-hr!66

+++ src/service/PerformanceService.ts
@@ -1,7 +1,9 @@
 'use strict';
-import { Comment, IComment, IPerformancePlan, IPerformanceReview, IPerformanceSection, PerformancePlan, PerformanceReview, PerformanceSection } from '../model/models'
+import { Comment, IComment, IPerformancePlan, IPerformanceReview, IPerformanceSection, PerformancePlan, PerformanceReview, PerformanceSection, IEmployee } from '../model/models'
 import { JABCError, JABCSuccess, JABCResponse } from '../utils/ResponseManager'
+import { Auth, getManagersByEmployee } from '../service/EmployeeService'
 import Database from '../database/Database';
+import IDatabaseClient from '../database/IDatabaseClient';
 
 /**
  * creates a new Comment for the PerformanceReview with [id]
@@ -14,11 +16,13 @@ import Database from '../database/Database';
  **/
 export async function createComment(id: Number, comment: IComment, xAuthToken: string) {
     try {
+        const performanceReview = await getPerformanceReview(id, xAuthToken)
+        const client = (await Auth(xAuthToken)).employee
         let res = await Database.getInstance().query('CALL create_performance_review_comment(?,?,?,?)', [
             id,
             comment.comment,
             comment.date,
-            comment.fkCommenter,
+            client.id,
         ], JABCResponse.PERFORMANCE)
 		return new JABCSuccess(JABCResponse.PERFORMANCE, `The comment was created successfully`)
     } catch (error) {
@@ -38,6 +42,11 @@ export async function createComment(id: Number, comment: IComment, xAuthToken: s
  **/
 export async function deleteComment(id: Number, idComment: Number, xAuthToken: string) {
     try {
+        const comment_ = await getComment(id, idComment, xAuthToken)
+        const client = (await Auth(xAuthToken)).employee
+        if(comment_.fkCommenter != client.id && client.adminLevel != IEmployee.adminLevelEnum.HR_ADMIN){
+            throw new JABCError(JABCResponse.EMPLOYEE, 'Can not edit a comment that is not yours.')
+        }
         let res = await Database.getInstance().query('CALL delete_performance_comment(?)', [
             idComment
         ], JABCResponse.PERFORMANCE)
@@ -58,6 +67,7 @@ export async function deleteComment(id: Number, idComment: Number, xAuthToken: s
  **/
 export async function deletePerformancePlan(id: Number, xAuthToken: string) {
     try {
+        const performancePlan = await getPerformancePlan(id, xAuthToken)
 		await Database.getInstance().query('CALL delete_performance_plan(?)', [id], JABCResponse.PERFORMANCE)
 		return new JABCSuccess(JABCResponse.EMPLOYEE, 'The performance plan was successfully deleted')
 	} catch (error) {
@@ -76,6 +86,7 @@ export async function deletePerformancePlan(id: Number, xAuthToken: string) {
  **/
 export async function deletePerformanceReview(id: Number, xAuthToken: string) {
     try {
+        const performanceReview = await getPerformanceReview(id, xAuthToken)
 		await Database.getInstance().query('CALL delete_performance_review(?)', [id], JABCResponse.PERFORMANCE)
 		return new JABCSuccess(JABCResponse.EMPLOYEE, 'The performance review was successfully deleted')
 	} catch (error) {
@@ -95,8 +106,9 @@ export async function deletePerformanceReview(id: Number, xAuthToken: string) {
  **/
 export async function getComment(id: Number, idComment: Number, xAuthToken: string) {
     try {
+        const performanceReview = await getPerformanceReview(id, xAuthToken)
         let res = await Database.getInstance().query('CALL get_comment(?)', [idComment], JABCResponse.PERFORMANCE)
-		return new Comment(res[0][0][0])
+        return new Comment(res[0][0][0])
     } catch (error) {
         throw error;
     }
@@ -113,6 +125,7 @@ export async function getComment(id: Number, idComment: Number, xAuthToken: stri
  **/
 export async function getComments(id: Number, xAuthToken: string) {
     try {
+        const performanceReview = await getPerformanceReview(id, xAuthToken)
         let res = await Database.getInstance().query('CALL get_performance_review_comments(?)', [id], JABCResponse.PERFORMANCE)
 		return Comment.Comments(res[0][0])
     } catch (error) {
@@ -131,8 +144,24 @@ export async function getComments(id: Number, xAuthToken: string) {
  **/
 export async function getPerformancePlan(id: Number, xAuthToken: string) {
     try {
+        const client = (await Auth(xAuthToken)).employee
         let res = await Database.getInstance().query('CALL get_performance_plan(?)', [id], JABCResponse.PERFORMANCE)
         let performancePlan = new PerformancePlan(res[0][0][0])
+        if (client.adminLevel == IEmployee.adminLevelEnum.MANAGER) {
+			let managed = false;
+			let managersOfEmployee = await getManagersByEmployee(performancePlan.fkEmployee, xAuthToken)
+			for (let manager of managersOfEmployee) {
+				if (manager.id === client.id) {
+					managed = true;
+					break;
+				}
+			}
+			if (!managed) {
+				throw new JABCError(JABCResponse.EMPLOYEE, 'The employee is not under the managed employees of the manager.')
+			}
+		}else if(client.adminLevel == IEmployee.adminLevelEnum.STAFF && client.id != performancePlan.fkEmployee){
+            throw new JABCError(JABCResponse.EMPLOYEE, 'An employee with STAFF admin level can not manage other employee\'s performance plan.')
+        }
         performancePlan.sections = PerformanceSection.PerformanceSections(res[0][1])
         performancePlan.comments = Comment.Comments(res[0][2])
         return performancePlan
@@ -152,8 +181,24 @@ export async function getPerformancePlan(id: Number, xAuthToken: string) {
  **/
 export async function getPerformanceReview(id: Number, xAuthToken: string) {
     try {
+        const client = (await Auth(xAuthToken)).employee
         let res = await Database.getInstance().query('CALL get_performance_review(?)', [id], JABCResponse.PERFORMANCE)
         let performanceReview = new PerformanceReview(res[0][0][0])
+        if (client.adminLevel == IEmployee.adminLevelEnum.MANAGER) {
+			let managed = false;
+			let managersOfEmployee = await getManagersByEmployee(performanceReview.fkEmployee, xAuthToken)
+			for (let manager of managersOfEmployee) {
+				if (manager.id === client.id) {
+					managed = true;
+					break;
+				}
+			}
+			if (!managed) {
+				throw new JABCError(JABCResponse.EMPLOYEE, 'The employee is not under the managed employees of the manager.')
+			}
+		}else if(client.adminLevel == IEmployee.adminLevelEnum.STAFF && client.id != performanceReview.fkEmployee){
+            throw new JABCError(JABCResponse.EMPLOYEE, 'An employee with STAFF admin level can not manage other employee\'s performance review.')
+        }
         performanceReview.sections = PerformanceSection.PerformanceSections(res[0][1])
         performanceReview.comments = Comment.Comments(res[0][2])
         return performanceReview
@@ -175,11 +220,16 @@ export async function getPerformanceReview(id: Number, xAuthToken: string) {
  **/
 export async function updateComment(id: Number, idComment: Number, comment: IComment, xAuthToken: string) {
     try {
+        const comment_ = await getComment(id, idComment, xAuthToken)
+        const client = (await Auth(xAuthToken)).employee
+        if(comment_.fkCommenter != client.id && client.adminLevel != IEmployee.adminLevelEnum.HR_ADMIN){
+            throw new JABCError(JABCResponse.EMPLOYEE, 'Can not edit a comment that is not yours.')
+        }
         let res = await Database.getInstance().query('CALL update_comment(?,?,?,?)', [
             idComment,
             comment.comment,
             comment.date,
-            comment.fkCommenter,
+            client.id
         ], JABCResponse.PERFORMANCE)
 		return new JABCSuccess(JABCResponse.PERFORMANCE, `The comment was edited successfully`)
     } catch (error) {
@@ -198,37 +248,44 @@ export async function updateComment(id: Number, idComment: Number, comment: ICom
  * @returns {Promise<IApiResponse>}
  **/
 export async function updatePerformancePlan(id: Number, performancePlan: IPerformancePlan, xAuthToken: string) {
-    let db = Database.getInstance();
+    let db: IDatabaseClient;
+    let conn: any;
+	try {
+		const performancePlan_ = await getPerformancePlan(id, xAuthToken)
+        db = Database.getInstance();
+        conn = db.initConnection();
+        await db.beginTransaction(conn);
 
-    try {
-        await db.beginTransaction();
-
-		let res = await db.rawQuery('CALL update_performance_plan(?,?,?)', [
+		let res = await db.rawQuery(conn, 'CALL update_performance_plan(?,?,?)', [
 			id,
 			performancePlan.date,
 			performancePlan.status
         ], JABCResponse.PERFORMANCE);
 
-		await db.rawQuery('CALL delete_plan_sections(?)', [id], JABCResponse.PERFORMANCE);
+		await db.rawQuery(conn, 'CALL delete_plan_sections(?)', [id], JABCResponse.PERFORMANCE);
 
-        for (let section of performancePlan.sections) {
-            await db.rawQuery('CALL create_employee_performance_plan_section(?,?,?)', [
-                id,
-                section.data,
-                section.sectionName
-            ], JABCResponse.PERFORMANCE);
+        if (performancePlan.sections != null){
+            for (let section of performancePlan.sections) {
+                await db.rawQuery(conn, 'CALL create_employee_performance_plan_section(?,?,?)', [
+                    id,
+                    section.data,
+                    section.sectionName
+                ], JABCResponse.PERFORMANCE);
+            }
         }
 
-        await db.commit();
-		await db.closeConnection();
+        await db.commit(conn);
 		return new JABCSuccess(JABCResponse.PERFORMANCE, `The performance plan was updated successfully`);
 	} catch (error) {
 		try {
-			await db.rollback();
-			await db.closeConnection();
-		} catch (err) { }
+			await db.rollback(conn);
+        } catch (err) { }
 		throw error;
-	}
+	} finally {
+        try {
+			await db.closeConnection(conn);
+        } catch (err) { }
+    }
 }
 
 
@@ -242,35 +299,42 @@ export async function updatePerformancePlan(id: Number, performancePlan: IPerfor
  * @returns {Promise<IApiResponse>}
  **/
 export async function updatePerformanceReview(id: Number, performanceReview: IPerformanceReview, xAuthToken: string) {
-    let db = Database.getInstance();
+    let db: IDatabaseClient;
+    let conn: any;
+	try {
+		const performanceReview_ = await getPerformanceReview(id, xAuthToken)
+        db = Database.getInstance();
+        conn = db.initConnection();
+        await db.beginTransaction(conn);
 
-    try {
-        await db.beginTransaction();
-
-		let res = await db.rawQuery('CALL update_performance_review(?,?,?)', [
+		let res = await db.rawQuery(conn,'CALL update_performance_review(?,?,?)', [
 			id,
 			performanceReview.date,
 			performanceReview.status
         ], JABCResponse.PERFORMANCE);
 
-        await db.rawQuery('CALL delete_review_sections(?)', [id], JABCResponse.PERFORMANCE);
+        await db.rawQuery(conn, 'CALL delete_review_sections(?)', [id], JABCResponse.PERFORMANCE);
 
-        for (let section of performanceReview.sections) {
-            await db.rawQuery('CALL create_employee_performance_review_section(?,?,?)', [
-                id,
-                section.data,
-                section.sectionName
-            ], JABCResponse.PERFORMANCE);
+        if (performanceReview.sections != null){
+            for (let section of performanceReview.sections) {
+                await db.rawQuery(conn, 'CALL create_employee_performance_review_section(?,?,?)', [
+                    id,
+                    section.data,
+                    section.sectionName
+                ], JABCResponse.PERFORMANCE);
+            }
         }
 
-        await db.commit();
-		await db.closeConnection();
+        await db.commit(conn);
 		return new JABCSuccess(JABCResponse.PERFORMANCE, `The performance review was updated successfully`);
 	} catch (error) {
 		try {
-			await db.rollback();
-			await db.closeConnection();
-		} catch (err) { }
+			await db.rollback(conn);
+        } catch (err) { }
 		throw error;
-	}
+	} finally {
+        try {
+			await db.closeConnection(conn);
+        } catch (err) { }
+    }
 }
